requirements.txt
pandas
numpy
matplotlib
scikit-learn
seaborn
tensorflow-gpu
flask